Final submission
I have returned the main function back to how it was originally intended. This means I had to put my scanf for C, N, and M in the readCourses function. I also cleared out a lot of comments and added some others. but all printf functions have been removed except for the printHighestTotal function. This was submitted and the auto test got all green check marks.